requirements.txt
discord.py==2.5.2
Flask==3.1.0
python-dotenv==1.0.1
waitress==3.0.2
